+++ CHANGELOG.md
@@ -31,6 +31,9 @@ ChangeLog
   its second argument, and no longer receives seperate properties and
   resourcetype arguments.
 * `MKCOL` now integrates better with propertystorage plugins.
+* The zip release ships with [sabre/vobject 3.4.2][vobj],
+  [sabre/http 4.0.0-alpha1][http], [sabre/event 2.0.1][evnt],
+  [sabre/uri 1.0.0][uri] and [sabre/xml 0.4.3][xml].
 
 
 2.2.0-alpha3 (2015-02-25)